chore: Add cSpell configuration and update dependencies
This commit adds a cSpell configuration file to the project, which includes the word "executemany" to the list of allowed words. Additionally, the go.mod file has been updated to include the latest versions of the dependencies, including the indirect dependency on github.com/golang-jwt/jwt v3.2.2+incompatible and golang.org/x/time v0.5.0.